// core/src/generated/api/LinkMutations.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { LinkExpressionInput } from "./LinkExpressionInput";
import type { LinkInput } from "./LinkInput";

export type LinkMutations = { additions: Array<LinkInput>, removals: Array<LinkExpressionInput>, };